9592-refactor to move formattedTrialSessionDetailsForFilteredEligibleCases into its own file [skip ci]
