+++ web-client/src/presenter/computeds/formattedTrialSessionDetails.js
@@ -62,29 +62,3 @@ export const formattedTrialSessionDetails = (get, applicationContext) => {
 
   return formattedTrialSession;
 };
-
-export const formattedTrialSessionDetailsForEligibleCases = (
-  get,
-  applicationContext,
-) => {
-  const trialSessionDetails = formattedTrialSessionDetails(
-    get,
-    applicationContext,
-  );
-  const filter = get(
-    state.screenMetadata.eligibleCasesFilter.hybridSessionFilter,
-  );
-
-  trialSessionDetails.formattedEligibleCases =
-    trialSessionDetails.formattedEligibleCases.filter(eligibleCase => {
-      if (!filter) return true;
-      if (filter === 'Small') {
-        return eligibleCase.docketNumberSuffix === 'S';
-      }
-      if (filter === 'Regular') {
-        return eligibleCase.docketNumberSuffix !== 'S';
-      }
-    });
-
-  return trialSessionDetails;
-};
