+++ web-client/src/presenter/state.js
@@ -56,10 +56,8 @@ import { formattedDocument } from './computeds/formattedDocument';
 import { formattedMessageDetail } from './computeds/formattedMessageDetail';
 import { formattedMessages } from './computeds/formattedMessages';
 import { formattedPendingItems } from './computeds/formattedPendingItems';
-import {
-  formattedTrialSessionDetails,
-  formattedTrialSessionDetailsForEligibleCases,
-} from './computeds/formattedTrialSessionDetails';
+import { formattedTrialSessionDetails } from './computeds/formattedTrialSessionDetails';
+import { formattedTrialSessionDetailsForFilteredEligibleCases } from './computeds/formattedTrialSessionDetailsForFilteredEligibleCases';
 import { formattedTrialSessions } from './computeds/formattedTrialSessions';
 import { formattedWorkQueue } from './computeds/formattedWorkQueue';
 import { getConstants } from '../getConstants';
@@ -174,7 +172,7 @@ const helpers = {
   formattedOpenCases,
   formattedPendingItems,
   formattedTrialSessionDetails,
-  formattedTrialSessionDetailsForEligibleCases,
+  formattedTrialSessionDetailsForFilteredEligibleCases,
   formattedTrialSessions,
   formattedWorkQueue,
   getTrialCityName,
