+++ web-client/src/views/TrialSessionDetail/EligibleCases.jsx
@@ -7,7 +7,8 @@ import React from 'react';
 export const EligibleCases = connect(
   {
     formattedEligibleCases:
-      state.formattedTrialSessionDetailsForEligibleCases.formattedEligibleCases,
+      state.formattedTrialSessionDetailsForFilteredEligibleCases
+        .formattedEligibleCases,
     trialSessionDetailsHelper: state.trialSessionDetailsHelper,
     trialSessionId: state.trialSession.trialSessionId,
     updateQcCompleteForTrialSequence:
